+++ pysteps/nowcasts/steps.py
@@ -28,7 +28,7 @@ from pysteps.nowcasts.utils import (
 )
 from pysteps.postprocessing import probmatching
 from pysteps.timeseries import autoregression, correlation
-from pysteps.utils.check_norain import check_norain
+from pysteps.utils.check_norain import check_norain, check_previous_radar_obs
 
 try:
     import dask
@@ -1504,7 +1504,7 @@ def forecast(
     """
 
     # Check the input precip and ar_order to be consistent
-    # zero-precip in previous time steps has to be removed
+    # zero-precip/constant field in previous time steps has to be removed
     # (constant field causes autoregression to fail)
     precip, ar_order = check_previous_radar_obs(precip, ar_order)
 
@@ -1547,77 +1547,3 @@ def forecast(
     nowcaster.reset_states_and_params()
     # Call the appropriate methods within the class
     return forecast_steps_nowcast
-
-
-# TODO: Where does this piece of code best fit: in utils or inside the class?
-def check_previous_radar_obs(precip, ar_order):
-    """Check all radar time steps and remove zero precipitation time steps
-
-    Parameters
-    ----------
-    precip : array-like
-      Array of shape (ar_order+1,m,n) containing the input precipitation fields
-      ordered by timestamp from oldest to newest. The time steps between the
-      inputs are assumed to be regular.
-    ar_order : int
-      The order of the autoregressive model to use. Must be >= 1.
-
-    Returns
-    -------
-    precip : numpy array
-      Array of shape (ar_order+1,m,n) containing the modified array with
-      input precipitation fields ordered by timestamp from oldest to newest.
-      The time steps between the inputs are assumed to be regular.
-    ar_order : int
-      The order of the autoregressive model to use. Must be >= 1.
-      Adapted to match with precip.shape equal (ar_order+1,m,n).
-    """
-    if not precip.shape[0] >= 2:
-        raise ValueError(
-            "Wrong precip shape. The radar input must have at least 2 time steps."
-        )
-
-    # Check all time steps for zero-precip (constant field, minimum==maximum)
-    zero_precip = [np.nanmin(obs) == np.nanmax(obs) for obs in precip]
-    if zero_precip[-1] or ~np.any(zero_precip):
-        # Unchanged if no rain in latest time step -> will be processed as zero_precip_radar=True
-        # or Unchanged if all time steps contain rain
-        return precip, ar_order
-    elif zero_precip[-2]:
-        # This case means radar-observed rain in the latest but no rain in the 2nd latest time steps.
-        # Solution 1:
-        # Assume the precipitation means clutter / parasit echoes in this case.
-        # Treat it as a default zero-precip case, AR-2 model
-        precip = np.ones((3, precip.shape[1], precip.shape[2])) * np.nanmin(precip)
-        # Give a warning
-        print(
-            "\n[WARNING] Precip + no-precip cases in the 2 latest radar input time steps.\nCannot calculate autoregression. Set to zero-precip radar input.\n"
-        )
-        return precip, 2
-        # # Solution 2 (to be discussed):
-        # # Adjust the radar input precipitation (if possible)
-        # # try to use a previous time step
-        # if not np.all(zero_precip[:-2]):
-        #     # find latest non-zero precip
-        #     # ATTENTION: This changes the time between precip[-2] and precip[-1] from initial 5min to a longer period
-        #     print(
-        #         "[WARNING] Radar input time steps adapted and ar_order set to 1. Input delta time changed."
-        #     )
-        #     prev = np.arange(len(zero_precip[:-2]))[~np.array(zero_precip[:-2])][-1]
-        #     # Adjust the time between input time steps to match the modified precip array
-        #     ### read the deltatime between the radar time step from metadata in current_deltatime
-        #     ### new_deltatime = (precip.shape[0] - prev) * current_deltatime
-        #     ### code here to set new_deltatimes the radar time step delta time in metadata
-        #     return precip[[prev, -1]], 1
-        # raise ValueError(
-        #     "Precipitation in latest but no previous time step. Not possible to calculate autoregression."
-        # )
-    else:
-        # Keep the latest time steps that do all contain precip
-        precip = precip[np.max(np.arange(len(zero_precip))[zero_precip]) + 1 :].copy()
-        if precip.shape[0] - 1 < ar_order:
-            # Give a warning
-            print(
-                f"[WARNING] Radar input only with {precip.shape[0]} non-zero time steps and ar_order set to {precip.shape[0]-1}."
-            )
-        return precip, min(ar_order, precip.shape[0] - 1)
